backend/Techfesia2019/requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.0.2
django-rest-swagger==2.2.0
django-sendgrid-v5==0.8.0
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
djangorestframework-simplejwt==4.3.0
firebase-admin==2.17.0
future==0.17.1
google-api-core==1.11.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.21.1
httplib2==0.13.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
openapi-codec==1.3.2
Pillow==6.1.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
python-http-client==3.1.0
pytz==2019.1
requests==2.22.0
rsa==4.0
sendgrid==6.0.5
sendgrid-django==4.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
Pillow==6.1.0
